[front] cleaning old resources

+++ server/OMenteJovem.Api/Domain/Models/NftTransferEvent.cs
@@ -8,7 +8,7 @@ public record NftTransferEvent
     public string Transaction { get; set; }
     public string FromAddress { get; set; }
     public string ToAddress { get; set; }
-    public double EventTimestamp { get; set; }
+    public long EventTimestamp { get; set; }
     public string EventType { get; set; } = "transfer";
     public NftChain NftChain { get; set; } = NftChain.Ethereum;
     public string NftIdentifier { get; set; }
